% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_hyp_bay_se.R
\docType{data}
\name{lf_hyp_bay_se}
\alias{lf_hyp_bay_se}
\title{Summarized Experiment of RNA-seq Counts}
\format{
An object of class \code{SummarizedExperiment} with 28395 rows and 16 columns.
}
\usage{
lf_hyp_bay_se
}
\description{
RNA was harvested from proliferating lung fibroblasts after 72 h of treatment
with molidustat or 0.5\% oxygen, separately or together with DMSO as the
vehicle control. RNA was extracted using Qiagen RNAeasy kit and sequenced
at BGI.
}
\details{
Reads were aligned with \code{Rsubread} using release 37 of the
\code{GRCh38.primary_assembly.genome.fa.gz} and annotated with the internal \emph{hg38}
NCBI RefSeq database. Entrez gene IDs were mapped to HGNC symbols and gene
names using \code{biomaRt}. The sequencing results are available from the NIH short
read archive (SRA) and the analysis code is contained in the \code{data-raw}
directory of the package source.

Phenotype data can be accessed with \code{colData(lf_hyp_bay_se)}:

\describe{
\item{id}{sample id}
\item{experiment}{biological replicate samples}
\item{oxygen}{ambient oxygen during culture}
\item{treatment}{
\code{DMSO} = DMSO vehicle 0.1\% \cr
\code{BAY} = molidustat 10 μM}
}

Simple gene annotation information can be accessed with \code{rowData(lf_hyp_bay_se)}:

\describe{
\item{hgnc_symbol}{gene symbol}
\item{description}{gene name}
}
}
\keyword{datasets}
