% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsubread_alignment_summary.R
\docType{data}
\name{rsubread_alignment_summary}
\alias{rsubread_alignment_summary}
\title{Summary of Rsubread alignment}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 27 columns.
}
\usage{
rsubread_alignment_summary
}
\description{
Tabular summary of the number of the total reads, mapped reads, uniquely
mapped reads, multi-mapped reads, unmapped reads, and indels from the
Rsubread alignment. Also includes data from the number of reads successfully
aligned.
}
\keyword{datasets}
